+++ test/smokes/rails_best_practices/expectations.rb
@@ -286,12 +286,9 @@ s.add_test(
       links: %w[https://rails-bestpractices.com/posts/2010/07/24/move-code-into-model/],
       id: "RailsBestPractices::Reviews::MoveCodeIntoModelReview",
       path: "app/views/articles/show.html.erb",
-      location: nil,
+      location: { start_line: 1 },
       object: nil,
       git_blame_info: nil
     }
-  ],
-  warnings: [
-    { message: "Invalid `line_number` is output: \"call\". The line location in `app/views/articles/show.html.erb` is lost.", file: "app/views/articles/show.html.erb" }
   ]
 )